requirements.txt
streamlit==1.37.1
faster-whisper==1.0.3
torch>=2.1
sentence-transformers==3.0.1
chromadb==0.5.5
pydub==0.25.1
soundfile==0.12.1
webrtcvad==2.0.10
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
PyYAML==6.0.1
rouge-score==0.1.2
jiwer==3.0.3
requests==2.31.0
yagmail==0.15.293
# faster-whisper
groq
reportlab

